# docs/swagger.yaml
basePath: /
definitions:
  utils.Delivery:
    properties:
      address:
        type: string
      city:
        type: string
      email:
        type: string
      name:
        type: string
      phone:
        type: string
      region:
        type: string
      zip:
        type: string
    type: object
  utils.Items:
    properties:
      brand:
        type: string
      chrt_id:
        type: integer
      name:
        type: string
      nm_id:
        type: integer
      price:
        type: integer
      rid:
        type: string
      sale:
        type: integer
      size:
        type: string
      status:
        type: integer
      total_price:
        type: integer
      track_number:
        type: string
    type: object
  utils.Message:
    properties:
      customer_id:
        type: string
      date_created:
        type: string
      delivery:
        $ref: '#/definitions/utils.Delivery'
      delivery_service:
        type: string
      entry:
        type: string
      internal_signature:
        type: string
      items:
        items:
          $ref: '#/definitions/utils.Items'
        type: array
      locale:
        type: string
      oof_shard:
        type: string
      order_uid:
        type: string
      payment:
        $ref: '#/definitions/utils.Payment'
      shardkey:
        type: string
      sm_id:
        type: integer
      track_number:
        type: string
    type: object
  utils.Payment:
    properties:
      amount:
        type: integer
      bank:
        type: string
      currency:
        type: string
      custom_fee:
        type: integer
      delivery_cost:
        type: integer
      goods_total:
        type: integer
      payment_dt:
        type: integer
      provider:
        type: string
      request_id:
        type: string
      transaction:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Документация к первому проекту "Wb-Техношколы"
  title: go-demoservice (prod. Nikita Parfenov)
  version: "1.0"
paths:
  /order/{order_uid}:
    get:
      description: Возвращает информацию о заказе
      parameters:
      - description: Order ID
        in: path
        name: order_uid
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Message'
        "404":
          description: Info wasn't found
          schema:
            type: string
      summary: Получить заказ по UUID (order_uid)
swagger: "2.0"